+++ google/cloud/storage/async/object_responses.h
