+++ google/cloud/storage/async/reader.h
@@ -15,9 +15,9 @@
 #ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_ASYNC_READER_H
 #define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_ASYNC_READER_H
 
-#include "google/cloud/storage/async_object_responses.h"
-#include "google/cloud/storage/async_reader_connection.h"
-#include "google/cloud/storage/async_token.h"
+#include "google/cloud/storage/async/object_responses.h"
+#include "google/cloud/storage/async/reader_connection.h"
+#include "google/cloud/storage/async/token.h"
 #include "google/cloud/status_or.h"
 #include "google/cloud/version.h"
 #include <memory>
