+++ google/cloud/storage/async/token.h
