+++ google/cloud/storage/async/token_test.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "google/cloud/storage/async_token.h"
+#include "google/cloud/storage/async/token.h"
 #include <gmock/gmock.h>
 #include <type_traits>
