+++ google/cloud/storage/async/writer.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "google/cloud/storage/async_writer.h"
+#include "google/cloud/storage/async/writer.h"
 #include "google/cloud/internal/make_status.h"
 
 namespace google {
