+++ google/cloud/storage/mocks/mock_async_connection.h
@@ -15,7 +15,7 @@
 #ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_MOCKS_MOCK_ASYNC_CONNECTION_H
 #define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_MOCKS_MOCK_ASYNC_CONNECTION_H
 
-#include "google/cloud/storage/async_connection.h"
+#include "google/cloud/storage/async/connection.h"
 #include "google/cloud/storage/object_metadata.h"
 #include "google/cloud/options.h"
 #include "google/cloud/status.h"
