+++ google/cloud/storage/storage_client_grpc_unit_tests.bzl
@@ -17,9 +17,9 @@
 """Automatically generated unit tests list - DO NOT EDIT."""
 
 storage_client_grpc_unit_tests = [
-    "async_reader_test.cc",
-    "async_token_test.cc",
-    "async_writer_test.cc",
+    "async/reader_test.cc",
+    "async/token_test.cc",
+    "async/writer_test.cc",
     "grpc_plugin_test.cc",
     "internal/async/accumulate_read_object_test.cc",
     "internal/async/connection_impl_test.cc",
